test rebalance suite toward unit-level coverage
