+++ tests/unit/test_global_state_reset.py
@@ -1,4 +1,4 @@
-"""Tests for global state reset functions."""
+"""Unit tests for module-level state reset helpers."""
 
 import pytest
 
@@ -8,7 +8,6 @@ from gateway.core.database import reset_db
 
 
 def test_reset_config_clears_state() -> None:
-    """Test that reset_config clears the global config."""
     config = GatewayConfig(
         database_url="postgresql://localhost/test",
         master_key="test",
@@ -23,14 +22,8 @@ def test_reset_config_clears_state() -> None:
 
 
 def test_reset_db_allows_reinit() -> None:
-    """Test that reset_db clears state so init_db can be called again.
-
-    We can't fully test init_db without a database, but we can verify
-    reset_db doesn't raise and clears the module state.
-    """
     from core import database
 
-    # Verify the function exists and runs without error when nothing is initialized
     reset_db()
 
     assert database._engine is None
